Issue #5104: moved and renamed wrong test class names
