+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/filefilters/exclusionbeforeexecutionfilefilter/InputBeforeExecutionExclusionFileFilterIncorrectClass.java
